Launch UI from main file
