+++ BlockMaker/gui/main_window.py
@@ -1,0 +1,12 @@
+from PyQt6.QtWidgets import QMainWindow
+from .gui import Ui_MainWindow
+
+
+class MainWindow(QMainWindow):
+    def __init__(self):
+        super().__init__()
+        self.ui = Ui_MainWindow()
+        self.ui.setupUi(self) 
+
+
+
